dom/base/test/meta_viewport/test_meta_viewport_change_name.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>name attribute changes</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta id="viewport" name="initial-name" content="width=device-width, initial-scale=1">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <script type="application/javascript">
    "use strict";

    add_task(async function change_name_attribute() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      // There is no valid <meta name="viewport"> so that viewport info values
      // are the same as values in test_meta_viewport0.html.
      is(info.defaultZoom, 0.25,  "initial scale is clamped to the default mimumim scale");
      is(info.minZoom, 0.25,      "minimum scale defaults to the absolute minimum");
      is(info.maxZoom,     10,    "maximum scale defaults to the absolute maximum");
      is(info.width,       980,   "width is the default width");
      is(info.height,      588,   "height is proportional to displayHeight");
      is(info.autoSize,    false, "autoSize is disabled by default");
      is(info.allowZoom,   true,  "zooming is enabled by default");

      // Now it's a valid viewport.
      // eslint-disable-next-line no-undef
      viewport.setAttribute("name", "viewport");
      info = getViewportInfo(800, 480);
      is(info.defaultZoom, 1, "initial zoom is 1");
      is(info.width, 800, "width should be 800");
      is(info.height, 480, "height should be 480");

      // Now it's invalid again.
      // eslint-disable-next-line no-undef
      viewport.setAttribute("name", "other");
      info = getViewportInfo(800, 480);
      is(info.defaultZoom, 0.25,  "initial scale is clamped to the default mimumim scale");
      is(info.minZoom, 0.25,      "minimum scale defaults to the absolute minimum");
      is(info.maxZoom,     10,    "maximum scale defaults to the absolute maximum");
      is(info.width,       980,   "width is the default width");
      is(info.height,      588,   "height is proportional to displayHeight");
      is(info.autoSize,    false, "autoSize is disabled by default");
      is(info.allowZoom,   true,  "zooming is enabled by default");
    });
  </script>
</body>
</html>
